requirements.txt
Flask==2.3.2
pymongo==4.7.0
opencv-python==4.8.0.74
Pillow==9.5.0
python-dotenv==1.0.0
numpy==1.26.0
